🚚 Move irrlicht test main to gui folder
